Extract common functionality to trait for easier reuse, fix default row height for XLSX
