+++ src/Spout/Writer/XLSX/Manager/WorksheetManager.php
@@ -14,6 +14,7 @@ use Box\Spout\Writer\Common\Creator\InternalEntityFactory;
 use Box\Spout\Writer\Common\Entity\Options;
 use Box\Spout\Writer\Common\Entity\Worksheet;
 use Box\Spout\Writer\Common\Helper\CellHelper;
+use Box\Spout\Writer\Common\Manager\ManagesCellSize;
 use Box\Spout\Writer\Common\Manager\RowManager;
 use Box\Spout\Writer\Common\Manager\Style\StyleMerger;
 use Box\Spout\Writer\Common\Manager\WorksheetManagerInterface;
@@ -25,6 +26,8 @@ use Box\Spout\Writer\XLSX\Manager\Style\StyleManager;
  */
 class WorksheetManager implements WorksheetManagerInterface
 {
+    use ManagesCellSize;
+
     /**
      * Maximum number of characters a cell can contain
      * @see https://support.office.com/en-us/article/Excel-specifications-and-limits-16c69c74-3d6a-4aaf-ba35-e6eb276e8eaa [Excel 2007]
@@ -62,17 +65,8 @@ EOD;
     /** @var InternalEntityFactory Factory to create entities */
     private $entityFactory;
 
-    /** @var float|null The default column width to use */
-    private $defaultColumnWidth;
-
-    /** @var float|null The default row height to use */
-    private $defaultRowHeight;
-
     /** @var bool Whether rows have been written */
     private $hasWrittenRows = false;
-
-    /** @var array Array of min-max-width arrays */
-    private $columnWidths;
 
     /**
      * WorksheetManager constructor.
@@ -115,54 +109,6 @@ EOD;
     public function getSharedStringsManager()
     {
         return $this->sharedStringsManager;
-    }
-
-    /**
-     * @param float|null $width
-     */
-    public function setDefaultColumnWidth($width)
-    {
-        $this->defaultColumnWidth = $width;
-    }
-
-    /**
-     * @param float|null $height
-     */
-    public function setDefaultRowHeight($height)
-    {
-        $this->defaultRowHeight = $height;
-    }
-
-    /**
-     * @param float $width
-     * @param array $columns One or more columns with this width
-     */
-    public function setColumnWidth(float $width, ...$columns)
-    {
-        // Gather sequences
-        $sequence = [];
-        foreach ($columns as $i) {
-            $sequenceLength = count($sequence);
-            if ($sequenceLength > 0) {
-                $previousValue = $sequence[$sequenceLength - 1];
-                if ($i !== $previousValue + 1) {
-                    $this->setColumnWidthForRange($width, $sequence[0], $previousValue);
-                    $sequence = [];
-                }
-            }
-            $sequence[] = $i;
-        }
-        $this->setColumnWidthForRange($width, $sequence[0], $sequence[count($sequence) - 1]);
-    }
-
-    /**
-     * @param float $width The width to set
-     * @param int $start First column index of the range
-     * @param int $end Last column index of the range
-     */
-    public function setColumnWidthForRange(float $width, int $start, int $end)
-    {
-        $this->columnWidths[] = [$start, $end, $width];
     }
 
     /**
@@ -226,7 +172,8 @@ EOD;
         $rowIndex = $worksheet->getLastWrittenRowIndex() + 1;
         $numCells = $row->getNumCells();
 
-        $rowXML = '<row r="' . $rowIndex . '" spans="1:' . $numCells . '">';
+        $hasCustomHeight = $this->defaultRowHeight > 0 ? '1' : '0';
+        $rowXML = "<row r=\"{$rowIndex}\" spans=\"1:{$numCells}\" customHeight=\"{$hasCustomHeight}\">";
 
         foreach ($row->getCells() as $cell) {
             $rowXML .= $this->applyStyleAndGetCellXML($cell, $rowStyle, $rowIndex, $cellIndex);
